Fix typo in load_kwargs dictionary key for dtype in prepare_model_and_processor_qwen3vl function

+++ src/drivevlms/preparation/qwen3_preparation.py
@@ -27,7 +27,7 @@ def prepare_model_and_processor_qwen3vl(config):
 
     load_kwargs = dict(
         trust_remote_code=True,
-        torch_dtype=config.dtype,
+        dtype=config.dtype,
     )
 
     if config.quantization:
